Abstract World class for evaluating embodied agents in isolation
Embodied here implies presence of sensors and motors that interact
with the rest of the world. Abstract base classes for these are
provided.

Test world for these is provided.

The code compiles, but hasn't been tested yet. Compilation required
a tweak in mbuild.py which is included into the commit.

	new file:   World/AbstractIsolatedEmbodiedWorld.cpp
	new file:   World/AbstractIsolatedEmbodiedWorld.h
	new file:   World/AbstractMotors.h
	new file:   World/AbstractSensors.h
	new file:   World/TestIsolatedEmbodiedWorld/TestIsolatedEmbodiedWorld.cpp
	new file:   World/TestIsolatedEmbodiedWorld/TestIsolatedEmbodiedWorld.h
	new file:   World/TestIsolatedEmbodiedWorld/TestMotors.h
	new file:   World/TestIsolatedEmbodiedWorld/TestSensors.h
	modified:   buildOptions.txt
	modified:   pythonTools/mbuild.py