No more	writable input nodes
MABE's Markov Brain implementation allowed gates to write into
input nodes. If the number of inputs is	high enough all	gates just
connect inputs to inputs and outputs are almost	alwas zero. The
problem	is made	worse by the fact that the system just silently
works if there are not enough bits in the genome to encode the
full range of possible inputs and/or outputs.

This commit implements two changes:
1. Input nodes can no longe be written into.
2. If node address fields in the genome can't take enough different
   values, a warning message is displayed.

	modified:   Brain/MarkovBrain/Gate/AbstractGate.cpp
	modified:   Brain/MarkovBrain/Gate/AbstractGate.h
	modified:   Brain/MarkovBrain/Gate/DecomposableFeedbackGate.cpp
	modified:   Brain/MarkovBrain/Gate/DecomposableFeedbackGate.h
	modified:   Brain/MarkovBrain/Gate/FeedbackGate.cpp
	modified:   Brain/MarkovBrain/Gate/FeedbackGate.h
	modified:   Brain/MarkovBrain/Gate/NeuronGate.h
	modified:   Brain/MarkovBrain/MarkovBrain.cpp
	modified:   Brain/MarkovBrain/MarkovBrain.h