Caching access to asteroid descriptions
AsteroidsDatasetParser now has a method emnabling access to
asteroid descriptions by asteroid name, bypassing the need to
query the filesystem every time the description is needed.

This introduces some minimal knowledge about the structure of the
descriptions into AsteroidDatasetParser, but substantially
decreases the time evaluations spend waiting for disk IO.

The remaining (and forgotten) issue is concurrent access to the
cache. In particular, if two cache misses happen simultaneously
there is a possibility of memory corruption. I did not encounter
this issue so far because operator[] is fast on maps, but that
can turn into something that is very difficult to debug.

	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeMentalImage.cpp
	modified:   World/AsteroidGazingWorld/Utilities/AsteroidsDatasetParser.cpp
	modified:   World/AsteroidGazingWorld/Utilities/AsteroidsDatasetParser.h