Seemingly working implementation of MNIST recognizer
The implementation compiles, which is promising in itself:
command data types are changed everywhere to std::tuple<unsigned>.
It was checked by valgrind for uninitialized variable dependencies
and there are none. I ran a single brain and the report by the
evaluator looks reasonable.

But who knows what else I could forget to change in this haunted
mansion-like codebase?

	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeFullMentalImage.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeFullMentalImage.h
	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeMentalImage.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeMentalImage.h
	modified:   World/AsteroidGazingWorld/MentalImages/commandLogger.h
	modified:   World/AsteroidGazingWorld/Sensors/AbsoluteFocusingSaccadingEyesSensors.cpp