Measures of saccading activity is available for optimization
The two new evaluations added to each Organism's dataMap upon
evaluation are sensorActivity and tieredSensorActivity.
The first one is the average number of saccades the brain has done
over the entire dataset, normalized by the length of the run and
number of asteroids to be in [0,1]. The second one is an integer
(stored as double) that shows which one on ten tiers the
sensorActivity value is at (number of tiers is tunable in code at
SpikesOnCubeMentalImage::evaluateOrganism).

	modified:   World/AsteroidGazingWorld/AsteroidGazingWorld.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeFullMentalImage.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeFullMentalImage.h
	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeMentalImage.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/SpikesOnCubeMentalImage.h
	modified:   World/AsteroidGazingWorld/Sensors/AbsoluteFocusingSaccadingEyesSensors.cpp
	modified:   World/AsteroidGazingWorld/Sensors/AbsoluteFocusingSaccadingEyesSensors.h