Brain JSONs instead of JSON strings
Organism's getJSONRecord() was simply forwarding the serialization
of the Brain into the corresponding JSON fields. This was creating
the situation when JSON has JSON-string valued fields, with all the
serialization and deserialization awkwardness that comes along with
that. Since all the brains I work with are JSON-describable, I've
made the function to deserialize the Brain strings before assigning
the fields.

	modified:   Organism/Organism.cpp