Publishing world state handles as global variables
In particular Im making global the string that is describing what
"slide" an AbstractSlideshowWorld is showing and the integer that
shows how many ticks has passed since the slide was first presented.
The machinery that existed for distributing this data over the
objects (sensors, mental images etc) is still in place, but it
should be removed in any future code cleanups.

	modified:   Global.cpp
	modified:   Global.h
	modified:   World/AbstractSlideshowWorld.h
	modified:   World/AsteroidGazingWorld/Schedules/AsteroidGazingSchedules.h