Pseudo-memory-leak problem solved
After getting rid of the old MTree memory leak I was able to locate
the code that allocated new memory at every generation. It
preserved the pointers to the memory somewhere, so valgrind's
memcheck couldn't find it. Through the use of massif and some trial
and error I determined that the problem disappears if I prevent the
constructor of LayeredBrain from reallocating its local
ParametersTable-s at every call (the ones it made were all the same
within each run anyway). So, the issue is still hidden, but now I
know that it can be avoided if no ParametersTable-s are reallocated
repeatedly. That will suffice for now.

	modified:   Brain/LayeredBrain/LayeredBrain.cpp
	modified:   Brain/LayeredBrain/LayeredBrain.h