Fix for the four-bit alignment bug
The four bit optimization was originally introduced for to cut
runtime, but it required alignment of bit ranges to the size four
1d grid. The alignment requirement has since been lifted, but there
was a forgotten call that assumed it. Since replacing that call with
a non-alignment-dependent version would lead it making a complicated
(read - unmaintainable) function, I decided to remove it altogether.

As a result, DistancesMentalImage now preserves the entire array
of double-encoded bits that gets passed to it at every timestep.
Might be slower, but at least it'll be easy to assure the correctness.

	modified:   World/AsteroidGazingWorld/MentalImages/DistancesMentalImage.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/DistancesMentalImage.h