LayeredBrain modified for sensorymotor awareness
If the sensors are active the brain must provide additional
outputs. In usual brains this happens automatically, but
LayeredBrain must decide where in its topology should these
outputs be. This modification makes it route additional outputs
from the first layer. For passive sensors the behavior should be
exactly the same as before.

	modified:   Brain/LayeredBrain/LayeredBrain.cpp
	modified:   Brain/LayeredBrain/LayeredBrain.h