Normalized, weighted distances
I have a suspicion that the failure of the distance-based approaches
is largely due to the fact that if the distance is measured for
the whole state, then changes in lower layers (very easy to
make) outweight changes in upper layers (hard to make). This commit
normalizes the contribution of each layer by the number of its units
and weights the results together, giving higher priority to upper
layers, to produce a final measure of distance between states.
Weighting is the same for intra- and cross-label distances, as the
magnitude of their importance for optimization varies across layers
in the same way, with the only difference being that the former is
minimized and the latter is maximized.

	modified:   World/AsteroidGazingWorld/MentalImages/DistancesMentalImage.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/DistancesMentalImage.h