New facility for passing data from sensors to brains
Due to transition to spatially aware perception I needed to pass
arbitrary data (more concretely, a boost::multi_array) between
the components of the perception pipeline. The idea to pass a void*
pointer as two doubles failed, as it would require recomputing the
addresses of inputs in the brain code too often. Instead I added
an attachToSensors(void*) method to AbstractBrain, a complimentary
void* getDataForBrain() method to AbstractSensors and the logical
call to these to AbstractIsolatedEmbodiedWorld that is made before
each evaluation.

	modified:   Brain/AbstractBrain.h
	deleted:    Utilities/doublepointer.h
	modified:   World/AbstractIsolatedEmbodiedWorld.cpp
	modified:   World/AbstractSensors.h