Gates with shift support
Now gates preserve IO nodes coordinates within filters and the
filter shifts separately. This is required because occasionally
(DETextureBrain::addCopyOfRandomlySelectedGate()) I need to add
a new gate just like a given one, but in a different filter. It
introduces an overhead due to the need to recompute absolute
coodinates of IO nodes each time a gate is attached to a new
Texture, however it is minimal since such attachment only happens
once per minibatch.

	modified:   Brain/DETextureBrain/Gate/AbstractTextureGate.h
	modified:   Brain/DETextureBrain/Gate/DeterministicTextureGate.cpp
	modified:   Brain/DETextureBrain/Gate/DeterministicTextureGate.h