Erasures counter + some useful messages at DETextureBrain
Adding a counter for bit erasures performed by gates has turned out
to be quite a feat. Since large overhead is not acceptable, I had
to distribute a pointer to the counter to every gate and make sure
its updated when the brain (which owns the variable) is copied or
deserialized. The result is valgrinded in both visualize and run
modes.

Additionally, some debug messages unrelated that were used to
get the idea that I need to count the erasures in the first place
got caught up in the commit. I may need them in the future, so I'm
preserving them commented out.

	modified:   Brain/AbstractBrain.h
	modified:   Brain/DETextureBrain/DETextureBrain.cpp
	modified:   Brain/DETextureBrain/DETextureBrain.h
	modified:   Brain/DETextureBrain/Gate/AbstractTextureGate.h
	modified:   Brain/DETextureBrain/Gate/DeterministicTextureGate.cpp
	modified:   Brain/DETextureBrain/Gate/DeterministicTextureGate.h
	modified:   Brain/DETextureBrain/Gate/ProbabilisticTextureGate.cpp
	modified:   Brain/DETextureBrain/Gate/ProbabilisticTextureGate.h
	modified:   Brain/TextureLayeredBrain/TextureLayeredBrain.cpp
	modified:   Brain/TextureLayeredBrain/TextureLayeredBrain.h
	modified:   World/AsteroidGazingWorld/MentalImages/ShapeMentalImage.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/ShapeMentalImage.h