Facilities for exact match counting
For now only the population encoding is implemented, but any fixed
encoding can be added just by swapping the matching function. The
code that does the counting is commented out for now because it
constrains the number of output to a fixed value and match counts
are irrelevant for most experiments.

	modified:   World/AsteroidGazingWorld/MentalImages/CompressedMentalImage.cpp
	modified:   World/AsteroidGazingWorld/MentalImages/CompressedMentalImage.h